third_party/blink/web_tests/virtual/controls-refresh/datetimelocal-picker/datetimelocal-hotkeys.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="datetime-local" id="datetime-local-0" value="2019-02-14T13:02">
<input type="datetime-local" id="datetime-local-1" value="2019-02-14T13:02">
<input type="datetime-local" id="datetime-local-2" value="2019-02-14T13:02">
<input type="datetime-local" id="datetime-local-3" value="2019-02-14T13:02">
<script>

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-0');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    eventSender.keyDown('t');

    let actualTodayDateString = new Date(dateTimeElement.value).toDateString();
    let expectedTodayDateString = new Date().toDateString();

    assert_equals(actualTodayDateString, expectedTodayDateString, "Today hotkey should have updated in-page control to today's date");
    
    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Datetimelocal picker: 't' key should select today's date and close popup");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-1');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    eventSender.keyDown('m');
    assert_equals(dateTimeElement.value, "2019-03-14T13:02", "Next month hotkey should have incremented month");

    eventSender.keyDown('M');
    assert_equals(dateTimeElement.value, "2019-02-14T13:02", "Previous month hotkey should have decremented month");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Datetimelocal picker: Previous/next month hotkey (m/M)");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-1');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    eventSender.keyDown('PageDown');
    assert_equals(dateTimeElement.value, "2019-03-14T13:02", "Next month hotkey should have incremented month");

    eventSender.keyDown('PageUp');
    assert_equals(dateTimeElement.value, "2019-02-14T13:02", "Previous month hotkey should have decremented month");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Datetimelocal picker: Previous/next month hotkey (PageDown/PageUp)");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-2');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    eventSender.keyDown('y');
    assert_equals(dateTimeElement.value, "2020-02-14T13:02", "Next year hotkey should have incremented year");

    eventSender.keyDown('Y');
    assert_equals(dateTimeElement.value, "2019-02-14T13:02", "Previous year hotkey should have decremented year");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Datetimelocal picker: Previous/next year hotkey");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-3');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    eventSender.keyDown('d');
    assert_equals(dateTimeElement.value, "2029-02-14T13:02", "Next year hotkey should have incremented decade");

    eventSender.keyDown('D');
    assert_equals(dateTimeElement.value, "2019-02-14T13:02", "Previous year hotkey should have decremented decade");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Datetimelocal picker: Previous/next decade hotkey");

</script>
</body>
</html>
